FrameExtractor now only pick unique frames + Cover with tests + Add comments
Before, one may randomly select frame indexes with replacement so it resulted in a frame labeled multiple times in ~labels.csv but only one frame image found in folder. So it creates inconsistent count of frames registered and frames effectively available as images. Now the user get the count of frames he asked for, I think this is an expected behaviour. It raises exception when inconsistent number of frames is found. In the future, on may relax this by maybe skipping the state that has not enough frames to cover the user needs, but still provide expected number of frames per state